Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
pydantic
python-dotenv
openpyxl
pandas
cryptography
multipart
python-multipart